Add model.js
