+++ server/gql/models/students/index.js
@@ -1,35 +1,16 @@
-import { student } from './query';
+import { GraphQLStudent } from './model';
 import { studentConnection } from './list';
-import { GraphQLInt, GraphQLNonNull } from 'graphql';
-import { limitAndOffset } from '@server/gql/fields/limitAndOffset';
-import db from '@server/database/models';
+import { StudentQueries } from './query';
 import { studentMutation } from './mutation';
 
-export const Student = student;
+// exporting graphql model of student.
+export const Student = GraphQLStudent;
 
 // exporting student connection.
 export const StudentConnection = studentConnection;
 
-export const studentQueries = {
-  args: {
-    id: {
-      type: new GraphQLNonNull(GraphQLInt)
-    }
-  },
-  query: {
-    type: student
-  },
-  list: {
-    ...StudentConnection,
-    resolve: StudentConnection.resolve,
-    type: StudentConnection.connectionType,
-    args: {
-      ...StudentConnection.connectionArgs,
-      ...limitAndOffset
-    }
-  },
-  model: db.students
-};
+// exporting student queries.
+export const studentQueries = StudentQueries;
 
 // exporting mutations on the students table.
 export const studentMutations = studentMutation;
