+++ server/gql/models/students/mutation.js
@@ -1,6 +1,6 @@
-import db from '@database/models';
 import { GraphQLID, GraphQLNonNull, GraphQLString } from 'graphql';
-import { student } from './query';
+import { GraphQLStudent } from './model';
+import db from '@database/models';
 import { customCreateResolver } from './customCreateResolver';
 
 export const studentsMutationFields = {
@@ -11,7 +11,7 @@ export const studentsMutationFields = {
 
 export const studentMutation = {
   args: studentsMutationFields,
-  type: student,
+  type: GraphQLStudent,
   model: db.students,
   customCreateResolver
 };
